４ステージ分の音楽を用意。 ステージセレクト画面でクリア時ヒットポイントを表示 入力が残るようなので、GameScreenのdispose時にInputProcessorをnullにしておく

+++ core/src/com/mygdx/game/Robo.java
@@ -210,7 +210,7 @@ public class Robo {
 		return _slowGauge;
 	}
 
-	public float getHitPoint() {
+	public int getHitPoint() {
 		return _hitPoint;
 	}
